Fixed bug with game not resetting upon death
Fixed bug where game was not resetting upon death.  Also, got the pause button to only render during game play and pause, but upon deployment of the gadme and death, only tap to play renders.

Now I am working on debugging the resetting of the apple count and hide the apple when user is prompted to tap the screen.  Everything else is working.