Pause button is now fully functioning
Got rid of the bugs resetting the apple count and position when pausing and not rendering and resetting correctly when the snake dies. 

This was acheived by making various revisions to the update, draw, drawPauseButton, and resetGame methods.  Further refactoring needed but pause feature is fully functional and bug free